Move bin.common to bin.lib.common
