+++ tools/run_tests/xds_k8s_test_driver/bin/lib/common.py
@@ -1,4 +1,4 @@
-# Copyright 2020 gRPC authors.
+# Copyright 2023 gRPC authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,11 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import logging
-import signal
-
-from absl import app
-from absl import flags
+"""Common functionality for bin/ python helpers."""
 
 from framework import xds_flags
 from framework import xds_k8s_flags
