+++ tools/run_tests/xds_k8s_test_driver/bin/run_td_setup.py
@@ -1,4 +1,4 @@
-# Copyright 2023 gRPC authors.
+# Copyright 2020 gRPC authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
