+++ tools/run_tests/xds_k8s_test_driver/bin/run_test_server.py
@@ -17,7 +17,7 @@ import signal
 from absl import app
 from absl import flags
 
-from bin import common
+from bin.lib import common
 from framework import xds_flags
 from framework import xds_k8s_flags
 from framework.infrastructure import gcp
